Extract duplicate fawe clipboard entitiy methods to super class
