+++ worldedit-core/src/main/java/com/fastasyncworldedit/core/extent/clipboard/CPUOptimizedClipboard.java
@@ -4,27 +4,18 @@ import com.fastasyncworldedit.core.jnbt.streamer.IntValueReader;
 import com.fastasyncworldedit.core.math.IntTriple;
 import com.sk89q.jnbt.CompoundTag;
 import com.sk89q.jnbt.IntTag;
-import com.sk89q.jnbt.NBTUtils;
 import com.sk89q.jnbt.Tag;
-import com.sk89q.worldedit.entity.BaseEntity;
-import com.sk89q.worldedit.entity.Entity;
-import com.sk89q.worldedit.extent.clipboard.BlockArrayClipboard;
 import com.sk89q.worldedit.math.BlockVector3;
 import com.sk89q.worldedit.regions.Region;
-import com.sk89q.worldedit.util.Location;
 import com.sk89q.worldedit.world.biome.BiomeType;
 import com.sk89q.worldedit.world.block.BaseBlock;
 import com.sk89q.worldedit.world.block.BlockState;
 import com.sk89q.worldedit.world.block.BlockStateHolder;
 
-import javax.annotation.Nullable;
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
-import java.util.UUID;
 
 public class CPUOptimizedClipboard extends LinearClipboard {
 
@@ -209,35 +200,6 @@ public class CPUOptimizedClipboard extends LinearClipboard {
             setTile(index, block.getNbtData());
         }
         return true;
-    }
-
-    @Nullable
-    @Override
-    public Entity createEntity(Location location, BaseEntity entity) {
-        BlockArrayClipboard.ClipboardEntity ret = new BlockArrayClipboard.ClipboardEntity(location, entity);
-        entities.add(ret);
-        return ret;
-    }
-
-    @Nullable
-    @Override
-    public Entity createEntity(Location location, BaseEntity entity, UUID uuid) {
-        Map<String, Tag> map = new HashMap<>(entity.getNbtData().getValue());
-        NBTUtils.addUUIDToMap(map, uuid);
-        entity.setNbtData(new CompoundTag(map));
-        BlockArrayClipboard.ClipboardEntity ret = new BlockArrayClipboard.ClipboardEntity(location, entity);
-        entities.add(ret);
-        return ret;
-    }
-
-    @Override
-    public List<? extends Entity> getEntities() {
-        return new ArrayList<>(entities);
-    }
-
-    @Override
-    public void removeEntity(Entity entity) {
-        this.entities.remove(entity);
     }
 
 }
